Move command to workspace subcommand
