+++ package.json
@@ -48,12 +48,7 @@
     "bin": "oclif-example",
     "devPlugins": [
       "@oclif/plugin-help"
-    ],
-    "topics": {
-      "abtest": {
-        "description": "ABtest commands. Run 'vtex abtest' to see all subcommands"
-      }
-    }
+    ]
   },
   "repository": "vtex/cli-plugin-abtest",
   "scripts": {
