+++ src/commands/workspace/abtest/start.ts
@@ -1,4 +1,4 @@
-import testAbStart from '../../modules/abtest/start'
+import testAbStart from '../../../modules/abtest/start'
 import { CustomCommand } from 'vtex'
 
 export default class ABTestStart extends CustomCommand {
